requirements.txt
numpy>=1.22.3
pandas>=1.4.2
python-dateutil>=2.8.2
fredapi>=0.5.0
yfinance>=0.1.70
sqlalchemy>=1.4.32
requests>=2.27.1
bs4>=0.0.1
beautifulsoup4>=4.11.1